+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.3] - 2020-04-16
+
 ## [1.1.2] - 2020-04-16
 ### Fixed
 - `matched` context value not being recalculated after its invalidation.